hal/dx12: intercept debug output
