+++ wgpu-hal/src/dx12/mod.rs
@@ -12,11 +12,12 @@ mod command;
 mod conv;
 mod descriptor;
 mod device;
+mod instance;
 
 use parking_lot::Mutex;
 use std::{borrow::Cow, ptr, sync::Arc};
 use winapi::{
-    shared::{dxgi, dxgi1_2, dxgi1_4, dxgi1_6, dxgitype, windef, winerror},
+    shared::{dxgi, dxgi1_2, dxgi1_4, dxgitype, windef, winerror},
     um::{d3d12, synchapi, winbase, winnt},
     Interface as _,
 };
@@ -96,14 +97,6 @@ pub struct Instance {
     factory: native::Factory4,
     library: Arc<native::D3D12Lib>,
     lib_dxgi: native::DxgiLib,
-}
-
-impl Drop for Instance {
-    fn drop(&mut self) {
-        unsafe {
-            self.factory.destroy();
-        }
-    }
 }
 
 unsafe impl Send for Instance {}
@@ -232,159 +225,6 @@ pub struct Fence {
 
 unsafe impl Send for Fence {}
 unsafe impl Sync for Fence {}
-
-impl crate::Instance<Api> for Instance {
-    unsafe fn init(desc: &crate::InstanceDescriptor) -> Result<Self, crate::InstanceError> {
-        let lib_main = native::D3D12Lib::new().map_err(|_| crate::InstanceError)?;
-
-        let lib_dxgi = native::DxgiLib::new().map_err(|_| crate::InstanceError)?;
-        let mut factory_flags = native::FactoryCreationFlags::empty();
-
-        if desc.flags.contains(crate::InstanceFlags::VALIDATION) {
-            // Enable debug layer
-            match lib_main.get_debug_interface() {
-                Ok(pair) => match pair.to_result() {
-                    Ok(debug_controller) => {
-                        debug_controller.enable_layer();
-                        debug_controller.Release();
-                    }
-                    Err(err) => {
-                        log::warn!("Unable to enable D3D12 debug interface: {}", err);
-                    }
-                },
-                Err(err) => {
-                    log::warn!("Debug interface function for D3D12 not found: {:?}", err);
-                }
-            }
-
-            // The `DXGI_CREATE_FACTORY_DEBUG` flag is only allowed to be passed to
-            // `CreateDXGIFactory2` if the debug interface is actually available. So
-            // we check for whether it exists first.
-            match lib_dxgi.get_debug_interface1() {
-                Ok(pair) => match pair.to_result() {
-                    Ok(debug_controller) => {
-                        debug_controller.destroy();
-                        factory_flags |= native::FactoryCreationFlags::DEBUG;
-                    }
-                    Err(err) => {
-                        log::warn!("Unable to enable DXGI debug interface: {}", err);
-                    }
-                },
-                Err(err) => {
-                    log::warn!("Debug interface function for DXGI not found: {:?}", err);
-                }
-            }
-        }
-
-        // Create DXGI factory
-        let factory = match lib_dxgi.create_factory2(factory_flags) {
-            Ok(pair) => match pair.to_result() {
-                Ok(factory) => factory,
-                Err(err) => {
-                    log::warn!("Failed to create DXGI factory: {}", err);
-                    return Err(crate::InstanceError);
-                }
-            },
-            Err(err) => {
-                log::warn!("Factory creation function for DXGI not found: {:?}", err);
-                return Err(crate::InstanceError);
-            }
-        };
-
-        Ok(Self {
-            factory,
-            library: Arc::new(lib_main),
-            lib_dxgi,
-        })
-    }
-
-    unsafe fn create_surface(
-        &self,
-        has_handle: &impl raw_window_handle::HasRawWindowHandle,
-    ) -> Result<Surface, crate::InstanceError> {
-        match has_handle.raw_window_handle() {
-            raw_window_handle::RawWindowHandle::Windows(handle) => Ok(Surface {
-                factory: self.factory,
-                wnd_handle: handle.hwnd as *mut _,
-                swap_chain: None,
-            }),
-            _ => Err(crate::InstanceError),
-        }
-    }
-    unsafe fn destroy_surface(&self, _surface: Surface) {
-        // just drop
-    }
-
-    unsafe fn enumerate_adapters(&self) -> Vec<crate::ExposedAdapter<Api>> {
-        // Try to use high performance order by default (returns None on Windows < 1803)
-        let factory6 = match self.factory.cast::<dxgi1_6::IDXGIFactory6>().to_result() {
-            Ok(f6) => {
-                // It's okay to decrement the refcount here because we
-                // have another reference to the factory already owned by `self`.
-                f6.destroy();
-                Some(f6)
-            }
-            Err(err) => {
-                log::info!("Failed to cast DXGI to 1.6: {}", err);
-                None
-            }
-        };
-
-        // Enumerate adapters
-        let mut adapters = Vec::new();
-        for cur_index in 0.. {
-            let raw = match factory6 {
-                Some(factory) => {
-                    let mut adapter2 = native::WeakPtr::<dxgi1_2::IDXGIAdapter2>::null();
-                    let hr = factory.EnumAdapterByGpuPreference(
-                        cur_index,
-                        dxgi1_6::DXGI_GPU_PREFERENCE_HIGH_PERFORMANCE,
-                        &dxgi1_2::IDXGIAdapter2::uuidof(),
-                        adapter2.mut_void(),
-                    );
-
-                    if hr == winerror::DXGI_ERROR_NOT_FOUND {
-                        break;
-                    }
-                    if let Err(err) = hr.to_result() {
-                        log::error!("Failed enumerating adapters: {}", err);
-                        break;
-                    }
-
-                    adapter2
-                }
-                None => {
-                    let mut adapter1 = native::WeakPtr::<dxgi::IDXGIAdapter1>::null();
-                    let hr = self
-                        .factory
-                        .EnumAdapters1(cur_index, adapter1.mut_void() as *mut *mut _);
-
-                    if hr == winerror::DXGI_ERROR_NOT_FOUND {
-                        break;
-                    }
-                    if let Err(err) = hr.to_result() {
-                        log::error!("Failed enumerating adapters: {}", err);
-                        break;
-                    }
-
-                    match adapter1.cast::<dxgi1_2::IDXGIAdapter2>().to_result() {
-                        Ok(adapter2) => {
-                            adapter1.destroy();
-                            adapter2
-                        }
-                        Err(err) => {
-                            log::error!("Failed casting to Adapter2: {}", err);
-                            break;
-                        }
-                    }
-                }
-            };
-
-            adapters.extend(Adapter::expose(raw, &self.library));
-        }
-        adapters
-    }
-}
 
 impl SwapChain {
     unsafe fn release_resources(self) -> native::WeakPtr<dxgi1_4::IDXGISwapChain3> {
